:art: Improving structure / format of File feature
Improving structure / format of File feature
